Move dimensions2sizes to staticmethod of Component
